Put HTTP server in its own class.
